requirements.txt
pyspark==3.4.0
kafka-python==2.0.2
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0
pandas==1.5.3
matplotlib==3.7.1